MTL File parsing Mouse hotkeys support Material rendering (WIP) Wheat To Bread Production Fixed configuration files pathing Added OpenGL Debugging

+++ src/renderEngine/shaders/glsl/animatedGameObjectVertexShader.glsl
@@ -27,7 +27,7 @@ uniform bool useNormalMap;
 uniform bool isInstanced;
 uniform bool areTangentsOn;
 
-uniform float numberOfRows;
+uniform int numberOfRows;
 uniform vec2 offset;
 
 uniform vec4 plane;
