+++ src/renderEngine/shaders/glsl/vertexShader.glsl
@@ -18,7 +18,7 @@ uniform vec3 lightPosition[10];
 
 uniform float useFakeLighting;
 
-uniform float numberOfRows;
+uniform int numberOfRows;
 uniform vec2 offset;
 
 uniform vec4 plane;
